move some files into cocos/core/algorithm move cocos/core/easing, cocos/core/utils/aotob and cocos/core/utils/murmurhash2_gc to cocos/core/algorithm
