+++ cocos/2d/renderer/render-data.ts
@@ -29,7 +29,7 @@ import { Material } from '../../asset/assets/material';
 import { TextureBase } from '../../asset/assets/texture-base';
 import { Color } from '../../core/math';
 import { Pool, RecyclePool } from '../../core/memop';
-import { murmurhash2_32_gc } from '../../core/utils/murmurhash2_gc';
+import { murmurhash2_32_gc } from '../../core/algorithm/murmurhash2_gc';
 import { SpriteFrame } from '../assets/sprite-frame';
 import { UIRenderer } from '../framework/ui-renderer';
 import { StaticVBAccessor, StaticVBChunk } from './static-vb-accessor';
