+++ cocos/3d/assets/skeleton.ts
@@ -26,7 +26,7 @@
 import { ccclass, type, serializable } from 'cc.decorator';
 import { CCString } from '../../core/data/utils/attribute';
 import { Mat4 } from '../../core/math';
-import { murmurhash2_32_gc } from '../../core/utils/murmurhash2_gc';
+import { murmurhash2_32_gc } from '../../core/algorithm/murmurhash2_gc';
 import type { DataPoolManager } from '../skeletal-animation/data-pool-manager';
 import { Asset } from '../../asset/assets/asset';
 import { legacyCC } from '../../core/global-exports';
