+++ cocos/core/algorithm/murmurhash2_gc.ts
