--- cocos/core/easing/index.ts
@@ -1,3 +1,0 @@
-import * as easing from './easing';
-
-export { easing };
