+++ cocos/core/splash-screen.ts
@@ -24,7 +24,7 @@
  */
 
 import { EDITOR, PREVIEW } from 'internal:constants';
-import * as easing from './easing/easing';
+import * as easing from './algorithm/easing';
 import { Material } from '../asset/assets/material';
 import { clamp01 } from './math/utils';
 import {
