+++ cocos/gfx/base/input-assembler.ts
@@ -23,7 +23,7 @@
  THE SOFTWARE.
  */
 
-import { murmurhash2_32_gc } from '../../core/utils/murmurhash2_gc';
+import { murmurhash2_32_gc } from '../../core/algorithm/murmurhash2_gc';
 import { Buffer } from './buffer';
 import { Attribute, GFXObject, ObjectType, InputAssemblerInfo, DrawInfo } from './define';
